refactor: move Realtime channel lifecycle into realtime.ts
Relocate the channel-lifecycle cluster — TableEntry, subscribeToChanges,
syncTableSubscription, the catch-all constants, and the channel-topic
numbering (realtimeChannelTopicPrefix/nextChannelTopic) — from db.ts into
realtime.ts, which now owns the Realtime subscription lifecycle. db.ts
keeps the per-QueryClient registry and supabaseCollectionOptions and
imports syncTableSubscription and TableEntry from realtime.ts.

This makes the module dependency one-way (db.ts -> realtime.ts) and keeps
the mutually recursive subscribeToChanges/syncTableSubscription pair in a
single module instead of splitting them across a circular import.
